Move config module to separate file.
