+++ apollo-router/src/plugins/telemetry/mod.rs
@@ -4,23 +4,19 @@ use crate::subscriber::replace_layer;
 use apollo_router_core::{register_plugin, Plugin};
 use apollo_spaceport::server::ReportSpaceport;
 use futures::FutureExt;
+use opentelemetry::global;
 use opentelemetry::propagation::TextMapPropagator;
 use opentelemetry::sdk::propagation::{
     BaggagePropagator, TextMapCompositePropagator, TraceContextPropagator,
 };
 use opentelemetry::sdk::trace::Builder;
-use opentelemetry::sdk::Resource;
 use opentelemetry::trace::TracerProvider;
-use opentelemetry::{global, Array, KeyValue, Value};
-use schemars::JsonSchema;
-use serde::{Deserialize, Serialize};
-use std::borrow::Cow;
-use std::collections::BTreeMap;
 use std::error::Error;
 use std::fmt;
 use tower::BoxError;
 use url::Url;
 
+pub mod config;
 pub mod metrics;
 pub mod tracing;
 
@@ -28,215 +24,6 @@ pub struct Telemetry {
     config: config::Conf,
     tracer_provider: Option<opentelemetry::sdk::trace::TracerProvider>,
     spaceport_shutdown: Option<futures::channel::oneshot::Sender<()>>,
-}
-pub mod config {
-    use super::*;
-
-    #[derive(Clone, Default, Debug, Deserialize, JsonSchema)]
-    #[serde(deny_unknown_fields, rename_all = "snake_case")]
-    pub struct Conf {
-        #[allow(dead_code)]
-        pub metrics: Option<Metrics>,
-        pub tracing: Option<Tracing>,
-        pub apollo: Option<tracing::apollo::Config>,
-    }
-
-    #[derive(Clone, Default, Debug, Deserialize, JsonSchema)]
-    #[serde(deny_unknown_fields, rename_all = "snake_case")]
-    #[allow(dead_code)]
-    pub struct Metrics {
-        pub otlp: Option<metrics::otlp::Config>,
-        pub prometheus: Option<metrics::prometheus::Config>,
-        pub datadog: Option<metrics::datadog::Config>,
-        pub apollo: Option<metrics::apollo::Config>,
-    }
-
-    #[derive(Clone, Default, Debug, Deserialize, JsonSchema)]
-    #[serde(deny_unknown_fields, rename_all = "snake_case")]
-    pub struct Tracing {
-        pub propagation: Option<Propagation>,
-        pub trace_config: Option<Trace>,
-        pub otlp: Option<tracing::otlp::Config>,
-        pub jaeger: Option<tracing::jaeger::Config>,
-        pub zipkin: Option<tracing::zipkin::Config>,
-        pub datadog: Option<tracing::datadog::Config>,
-    }
-
-    #[derive(Clone, Default, Debug, Deserialize, JsonSchema)]
-    #[serde(deny_unknown_fields, rename_all = "snake_case")]
-    pub struct Propagation {
-        pub baggage: Option<bool>,
-        pub trace_context: Option<bool>,
-        pub jaeger: Option<bool>,
-        pub datadog: Option<bool>,
-        pub zipkin: Option<bool>,
-    }
-
-    #[derive(Default, Debug, Clone, Deserialize, Serialize, JsonSchema)]
-    #[serde(deny_unknown_fields)]
-    pub struct Trace {
-        pub service_name: Option<String>,
-        pub service_namespace: Option<String>,
-        pub sampler: Option<SamplerOption>,
-        pub parent_based_sampler: Option<bool>,
-        pub max_events_per_span: Option<u32>,
-        pub max_attributes_per_span: Option<u32>,
-        pub max_links_per_span: Option<u32>,
-        pub max_attributes_per_event: Option<u32>,
-        pub max_attributes_per_link: Option<u32>,
-        pub attributes: Option<BTreeMap<String, AttributeValue>>,
-    }
-
-    #[derive(Debug, Clone, Deserialize, Serialize, JsonSchema)]
-    #[serde(untagged, deny_unknown_fields)]
-    pub enum AttributeValue {
-        /// bool values
-        Bool(bool),
-        /// i64 values
-        I64(i64),
-        /// f64 values
-        F64(f64),
-        /// String values
-        String(String),
-        /// Array of homogeneous values
-        Array(AttributeArray),
-    }
-
-    impl From<AttributeValue> for opentelemetry::Value {
-        fn from(value: AttributeValue) -> Self {
-            match value {
-                AttributeValue::Bool(v) => Value::Bool(v),
-                AttributeValue::I64(v) => Value::I64(v),
-                AttributeValue::F64(v) => Value::F64(v),
-                AttributeValue::String(v) => Value::String(Cow::from(v)),
-                AttributeValue::Array(v) => Value::Array(v.into()),
-            }
-        }
-    }
-
-    #[derive(Debug, Clone, Deserialize, Serialize, JsonSchema)]
-    #[serde(untagged, deny_unknown_fields)]
-    pub enum AttributeArray {
-        /// Array of bools
-        Bool(Vec<bool>),
-        /// Array of integers
-        I64(Vec<i64>),
-        /// Array of floats
-        F64(Vec<f64>),
-        /// Array of strings
-        String(Vec<Cow<'static, str>>),
-    }
-
-    impl From<AttributeArray> for opentelemetry::Array {
-        fn from(array: AttributeArray) -> Self {
-            match array {
-                AttributeArray::Bool(v) => Array::Bool(v),
-                AttributeArray::I64(v) => Array::I64(v),
-                AttributeArray::F64(v) => Array::F64(v),
-                AttributeArray::String(v) => Array::String(v),
-            }
-        }
-    }
-
-    #[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
-    #[serde(deny_unknown_fields, untagged)]
-    pub enum SamplerOption {
-        /// Sample a given fraction of traces. Fractions >= 1 will always sample. If the parent span is
-        /// sampled, then it's child spans will automatically be sampled. Fractions < 0 are treated as
-        /// zero, but spans may still be sampled if their parent is.
-        TraceIdRatioBased(f64),
-        Always(Sampler),
-    }
-
-    #[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
-    #[serde(deny_unknown_fields)]
-    pub enum Sampler {
-        /// Always sample the trace
-        AlwaysOn,
-        /// Never sample the trace
-        AlwaysOff,
-    }
-
-    impl From<&Trace> for opentelemetry::sdk::trace::Config {
-        fn from(config: &Trace) -> Self {
-            let mut trace_config = opentelemetry::sdk::trace::config();
-
-            let sampler = match (&config.sampler, &config.parent_based_sampler) {
-                (Some(SamplerOption::Always(Sampler::AlwaysOn)), Some(true)) => {
-                    Some(parent_based(opentelemetry::sdk::trace::Sampler::AlwaysOn))
-                }
-                (Some(SamplerOption::Always(Sampler::AlwaysOff)), Some(true)) => {
-                    Some(parent_based(opentelemetry::sdk::trace::Sampler::AlwaysOff))
-                }
-                (Some(SamplerOption::TraceIdRatioBased(ratio)), Some(true)) => Some(parent_based(
-                    opentelemetry::sdk::trace::Sampler::TraceIdRatioBased(*ratio),
-                )),
-                (Some(SamplerOption::Always(Sampler::AlwaysOn)), _) => {
-                    Some(opentelemetry::sdk::trace::Sampler::AlwaysOn)
-                }
-                (Some(SamplerOption::Always(Sampler::AlwaysOff)), _) => {
-                    Some(opentelemetry::sdk::trace::Sampler::AlwaysOff)
-                }
-                (Some(SamplerOption::TraceIdRatioBased(ratio)), _) => Some(
-                    opentelemetry::sdk::trace::Sampler::TraceIdRatioBased(*ratio),
-                ),
-                (_, _) => None,
-            };
-            if let Some(sampler) = sampler {
-                trace_config = trace_config.with_sampler(sampler);
-            }
-            if let Some(n) = config.max_events_per_span {
-                trace_config = trace_config.with_max_events_per_span(n);
-            }
-            if let Some(n) = config.max_attributes_per_span {
-                trace_config = trace_config.with_max_attributes_per_span(n);
-            }
-            if let Some(n) = config.max_links_per_span {
-                trace_config = trace_config.with_max_links_per_span(n);
-            }
-            if let Some(n) = config.max_attributes_per_event {
-                trace_config = trace_config.with_max_attributes_per_event(n);
-            }
-            if let Some(n) = config.max_attributes_per_link {
-                trace_config = trace_config.with_max_attributes_per_link(n);
-            }
-
-            let mut resource_defaults = vec![];
-            if let Some(service_name) = &config.service_name {
-                resource_defaults.push(KeyValue::new(
-                    opentelemetry_semantic_conventions::resource::SERVICE_NAME,
-                    service_name.clone(),
-                ));
-            }
-            if let Some(service_namespace) = &config.service_namespace {
-                resource_defaults.push(KeyValue::new(
-                    opentelemetry_semantic_conventions::resource::SERVICE_NAMESPACE,
-                    service_namespace.clone(),
-                ));
-            }
-            let resource = Resource::new(resource_defaults).merge(&mut Resource::new(
-                config
-                    .attributes
-                    .clone()
-                    .unwrap_or_default()
-                    .iter()
-                    .map(|(k, v)| {
-                        KeyValue::new(
-                            opentelemetry::Key::from(k.clone()),
-                            opentelemetry::Value::from(v.clone()),
-                        )
-                    })
-                    .collect::<Vec<KeyValue>>(),
-            ));
-
-            trace_config = trace_config.with_resource(resource);
-            trace_config
-        }
-    }
-}
-
-fn parent_based(sampler: opentelemetry::sdk::trace::Sampler) -> opentelemetry::sdk::trace::Sampler {
-    opentelemetry::sdk::trace::Sampler::ParentBased(Box::new(sampler))
 }
 
 #[derive(Debug)]
